Test out fzf lua again
Has nicer compact ui.

It seems to have nicer global grep that uses fuzzy search instead of
just grep.

It has nicer way to not use gitignore with a keybind.

You can also switch between grep and fuzzy search easily.

It has a preview for code actions.

The only minus seems to be that it doesn't have syntax highlighting
when searching current buffer and symbol search is a little bit uglier.